fix(rubocop): extract report logging helpers
Issue: No ticket/issue
